Extract client method for filters in FilterClient class
