+++ tests/clients/BarracksClient.spec.js
@@ -43,6 +43,17 @@ describe('Barracks', () => {
       expect(barracksClient).to.have.property('setGoogleAnalyticsTrackingId').and.to.be.a('function');
     });
 
+    it('should initialize methods from filterClient when constructor called', () => {
+      // When
+      const barracksClient = new Barracks();
+
+      // Then
+      expect(barracksClient).to.have.property('createFilter').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getFilter').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getFilters').and.to.be.a('function');
+      expect(barracksClient).to.have.property('deleteFilter').and.to.be.a('function');
+    });
+
     it('should initialize methods from segmentClient when constructor called', () => {
       // When
       const barracksClient = new Barracks();
@@ -176,278 +187,6 @@ describe('Barracks', () => {
         expect(mockedBasename).to.have.been.calledOnce;
         expect(mockedBasename).to.have.been.calledWithExactly(file);
         done();
-      });
-    });
-  });
-
-  describe('#createFilter()', () => {
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const filter = { name: 'Filter', query: { eq: { unitId: 'value' } } };
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.createFilter(token, filter).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('createFilterV1', {
-          headers: { 'x-auth-token': token },
-          body: filter
-        });
-        done();
-      });
-    });
-
-    it('should return the created filter', done => {
-      // Given
-      const filter = { name: 'Filter', query: { eq: { unitId: 'value' } } };
-      const savedFilter = Object.assign({}, filter, { userId: '123456789' });
-      const response = { body: savedFilter };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.createFilter(token, filter).then(result => {
-        expect(result).to.be.equals(savedFilter);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('createFilterV1', {
-          headers: { 'x-auth-token': token },
-          body: filter
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-    it('should return an error message when request fails and v2 flag is enabled', done => {
-      // Given
-      barracks.v2Enabled = true;
-      const filter = { name: 'Filter', query: { eq: { unitId: 'value' } } };
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.createFilter(token, filter).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('createFilterV2', {
-          headers: { 'x-auth-token': token },
-          body: filter
-        });
-        done();
-      });
-    });
-
-    it('should return the created filter and v2 flag is enabled', done => {
-      // Given
-      barracks.v2Enabled = true;
-      const filter = { name: 'Filter', query: { eq: { unitId: 'value' } } };
-      const savedFilter = Object.assign({}, filter, { userId: '123456789' });
-      const response = { body: savedFilter };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.createFilter(token, filter).then(result => {
-        expect(result).to.be.equals(savedFilter);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('createFilterV2', {
-          headers: { 'x-auth-token': token },
-          body: filter
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#deleteFilter()', () => {
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const name = 'aFilter';
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.deleteFilter(token, name).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('deleteFilterV1', {
-          headers: { 'x-auth-token': token },
-          pathVariables: { filter: name }
-        });
-        done();
-      });
-    });
-
-    it('should return nothing when filter is deleted', done => {
-      // Given
-      const name = 'aFilter';
-      const response = { body: undefined };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.deleteFilter(token, name).then(result => {
-        expect(result).to.be.equals(undefined);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('deleteFilterV1', {
-          headers: { 'x-auth-token': token },
-          pathVariables: { filter: name }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-    it('should return an error message when request fails and v2 flag is enabled', done => {
-      // Given
-      barracks.v2Enabled = true;
-      const name = 'aFilter';
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.deleteFilter(token, name).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('deleteFilterV2', {
-          headers: { 'x-auth-token': token },
-          pathVariables: { filter: name }
-        });
-        done();
-      });
-    });
-
-    it('should return nothing when filter is deleted and v2 flag is enabled', done => {
-      // Given
-      barracks.v2Enabled = true;
-      const name = 'aFilter';
-      const response = { body: undefined };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.deleteFilter(token, name).then(result => {
-        expect(result).to.be.equals(undefined);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('deleteFilterV2', {
-          headers: { 'x-auth-token': token },
-          pathVariables: { filter: name }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getFilterByName()', () => {
-
-    const filterName = 'myCoolFilter';
-    const filter = { name: filterName, query: { eq: { unitId: 'plop' } } };
-
-    it('should return an error if v2Enabled and request fails', done => {
-      // Given
-      barracks.v2Enabled = true;
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getFilterByName(token, filterName).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getFilter', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            filter: filterName
-          }
-        });
-        done();
-      });
-    });
-
-    it('should return specified filter when v2Enabled and request succeeds', done => {
-      // Given
-      barracks.v2Enabled = true;
-      const response = { body: filter };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getFilterByName(token, filterName).then(result => {
-        expect(result).to.be.equals(filter);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getFilter', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            filter: filterName
-          }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getFilters()', () => {
-
-    it('should forward to the client with correct headers', done => {
-      // Given
-      const options = {
-        headers: { 'x-auth-token': token },
-      }
-      barracks.client.retrieveAllPages = sinon.spy();
-
-      // When / Then
-      barracks.getFilters(token).then(result => {
-        expect(result).to.be.instanceOf(PageableStream);
-        expect(barracks.client.retrieveAllPages).to.have.been.calledOnce;
-        expect(barracks.client.retrieveAllPages).to.have.been.calledWithExactly(
-          new PageableStream(),
-          'getFiltersV1',
-          options,
-          'filters'
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-    it('should forward to the client with correct headers when v2 flag is enabled', done => {
-      // Given
-      barracks.v2Enabled = true;
-      const options = {
-        headers: { 'x-auth-token': token },
-      }
-      barracks.client.retrieveAllPages = sinon.spy();
-
-      // When / Then
-      barracks.getFilters(token).then(result => {
-        expect(result).to.be.instanceOf(PageableStream);
-        expect(barracks.client.retrieveAllPages).to.have.been.calledOnce;
-        expect(barracks.client.retrieveAllPages).to.have.been.calledWithExactly(
-          new PageableStream(),
-          'getFiltersV2',
-          options,
-          'filters'
-        );
-        done();
-      }).catch(err => {
-        done(err);
       });
     });
   });
